dom/tests/mochitest/ajax/offline/test_cancelOfflineCache.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/simpleManifest.cacheManifest">
<head>
<title>Cancel offline cache</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

var manifest = "http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/simpleManifest.cacheManifest";
var manifestURI = Cc["@mozilla.org/network/io-service;1"]
                  .getService(Ci.nsIIOService)
                  .newURI(manifest, null, null);
var updateService = Cc['@mozilla.org/offlinecacheupdate-service;1']
                    .getService(Ci.nsIOfflineCacheUpdateService);

function manifestCached () {
  OfflineTest.ok(false, "The update was supposed to be canceled");

  OfflineTest.teardown();
  OfflineTest.finish();
}

function onError () {
  OfflineTest.ok(true, "Expected error: Update canceled");

  OfflineTest.teardown();
  OfflineTest.finish();
}

function onProgress () {
  var i = 0;
  while (i < updateService.numUpdates) {
    var update = updateService.getUpdate(i);
    if (update.manifestURI.spec == manifestURI.spec) {
      update.cancel();
      return;
    }
    i++;
  }
}

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.priv(onError);
  applicationCache.onprogress = OfflineTest.priv(onProgress);
  applicationCache.oncached = OfflineTest.priv(manifestCached);
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body>

</body>
</html>
